Re-implemented the display data struct
I had initially tried to make the display label part of the data struct.
This became confusing when trying to access the label in the builder
file through the template binds. Now the data is it's own struct and
leaves managing the label to the window, AS IT ALWAYS SHOULD HAVE BEEN.